Add client retry delay to unit tests

+++ internal/config/config.go
@@ -22,7 +22,7 @@ type Config struct {
 type Client struct {
 	SkipSSLVerification bool  `default:"false" envconfig:"CLIENT_SKIP_TLS_VERIFICATION"`
 	RetryDelay          int64 `default:"1" envconfig:"CLIENT_RETRY_DELAY_SECONDS"`
-	Timeout             uint  `default:"20" envconfig:"CLIENT_TIMEOUT_SECONDS"`
+	Timeout             int64 `default:"20" envconfig:"CLIENT_TIMEOUT_SECONDS"`
 }
 
 type GravitySettings struct {
